feat: add node test utils
