+++ crates/node/src/lib.rs
@@ -5,6 +5,8 @@ mod args;
 mod bridge;
 mod constants;
 mod node;
+#[cfg(feature = "test-utils")]
+pub mod test_utils;
 
 pub use args::*;
 pub use node::ScrollRollupNode;
