+++ crates/node/tests/e2e.rs
@@ -1,26 +1,13 @@
 //! End-to-end tests for the rollup node.
 
-use alloy_primitives::B256;
+use alloy_primitives::Signature;
 use futures::StreamExt;
-use reth_e2e_test_utils::{node::NodeTestContext, NodeHelperType};
 use reth_network::{NetworkConfigBuilder, PeersInfo};
-use reth_network_peers::PeerId;
-use reth_node_api::PayloadBuilderAttributes;
-use reth_node_builder::{NodeBuilder, NodeConfig, NodeHandle};
-use reth_node_core::args::{DiscoveryArgs, NetworkArgs, RpcServerArgs};
-use reth_rpc_server_types::RpcModuleSelection;
-use reth_scroll_chainspec::ScrollChainSpec;
-use reth_scroll_engine_primitives::ScrollPayloadBuilderAttributes;
+use reth_scroll_chainspec::SCROLL_DEV;
 use reth_scroll_node::ScrollNetworkPrimitives;
-use reth_tasks::TaskManager;
-use rollup_node::{
-    BeaconProviderArgs, L1ProviderArgs, L2ProviderArgs, NetworkArgs as ScrollNetworkArgs,
-    ScrollRollupNode, ScrollRollupNodeConfig, SequencerArgs,
-};
-use scroll_alloy_rpc_types_engine::ScrollPayloadAttributes;
-use scroll_network::{NewBlockWithPeer, SCROLL_MAINNET};
+use rollup_node::test_utils::build_bridge_node;
+use scroll_network::NewBlockWithPeer;
 use scroll_wire::ScrollWireConfig;
-use std::{path::PathBuf, sync::Arc};
 use tracing::trace;
 
 /// We test the bridge from the eth-wire protocol to the scroll-wire protocol.
@@ -37,20 +24,19 @@ use tracing::trace;
 async fn can_bridge_blocks() {
     reth_tracing::init_test_tracing();
 
-    // Create the chain spec for scroll mainnet with Darwin v2 activated and a test genesis.
-    let chain_spec = (*SCROLL_MAINNET).clone();
+    // Create the chain spec for scroll dev with Euclid v2 activated and a test genesis.
+    let chain_spec = (*SCROLL_DEV).clone();
 
     // Setup the bridge node and a standard node.
     let (mut bridge_node, tasks, bridge_peer_id) =
         build_bridge_node(chain_spec.clone()).await.expect("Failed to setup nodes");
 
     // Instantiate the scroll NetworkManager.
-    let network_config =
-        NetworkConfigBuilder::<reth_scroll_node::ScrollNetworkPrimitives>::with_rng_secret_key()
-            .disable_discovery()
-            .with_unused_listener_port()
-            .with_pow()
-            .build_with_noop_provider(chain_spec.clone());
+    let network_config = NetworkConfigBuilder::<ScrollNetworkPrimitives>::with_rng_secret_key()
+        .disable_discovery()
+        .with_unused_listener_port()
+        .with_pow()
+        .build_with_noop_provider(chain_spec.clone());
     let scroll_wire_config = ScrollWireConfig::new(true);
     let mut scroll_network =
         scroll_network::ScrollNetworkManager::new(network_config, scroll_wire_config).await;
@@ -80,12 +66,17 @@ async fn can_bridge_blocks() {
     bridge_node.network.add_peer(network_handle.local_node_record()).await;
     bridge_node.network.next_session_established().await;
 
-    tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
-
     // Send a block from the standard NetworkManager to the bridge node.
-    let block_1: reth_scroll_primitives::ScrollBlock =
-        serde_json::from_str(include_str!("../assets/block_1.json")).unwrap();
+    let mut block_1: reth_scroll_primitives::ScrollBlock =
+        serde_json::from_str(include_str!("../assets/block.json")).unwrap();
+
+    // Compute the block hash while masking the extra data which isn't used in block hash
+    // computation.
+    let extra_data = block_1.extra_data.clone();
+    block_1.header.extra_data = Default::default();
     let block_1_hash = block_1.hash_slow();
+    block_1.header.extra_data = extra_data.clone();
+
     let new_block_1 = reth_eth_wire_types::NewBlock { block: block_1, ..Default::default() };
 
     trace!("Announcing block to eth-wire network");
@@ -95,65 +86,17 @@ async fn can_bridge_blocks() {
     if let Some(scroll_network::NetworkManagerEvent::NewBlock(NewBlockWithPeer {
         peer_id,
         block,
-        signature: _,
+        signature,
     })) = scroll_network.next().await
     {
         assert_eq!(peer_id, bridge_peer_id);
         assert_eq!(block.hash_slow(), block_1_hash);
+        assert_eq!(
+            TryInto::<Signature>::try_into(extra_data.as_ref().windows(65).last().unwrap())
+                .unwrap(),
+            signature
+        )
     } else {
         panic!("Failed to receive block from scroll-wire network");
     }
-}
-
-// HELPERS
-// ---------------------------------------------------------------------------------------------
-/// Helper function to create a new bridge node that will bridge messages from the eth-wire
-pub async fn build_bridge_node(
-    chain_spec: Arc<ScrollChainSpec>,
-) -> eyre::Result<(NodeHelperType<ScrollRollupNode>, TaskManager, PeerId)> {
-    // Create a [`TaskManager`] to manage the tasks.
-    let tasks = TaskManager::current();
-    let exec = tasks.executor();
-
-    // Define the network configuration with discovery disabled.
-    let network_config = NetworkArgs {
-        discovery: DiscoveryArgs { disable_discovery: true, ..DiscoveryArgs::default() },
-        ..NetworkArgs::default()
-    };
-
-    // Create the node config
-    let node_config = NodeConfig::new(chain_spec.clone())
-        .with_network(network_config.clone())
-        .with_rpc(RpcServerArgs::default().with_http().with_http_api(RpcModuleSelection::All))
-        .set_dev(false);
-
-    // Create the node for a bridge node that will bridge messages from the eth-wire protocol
-    // to the scroll-wire protocol.
-    let node_args = ScrollRollupNodeConfig {
-        test: true,
-        network_args: ScrollNetworkArgs {
-            enable_eth_scroll_wire_bridge: true,
-            enable_scroll_wire: true,
-        },
-        optimistic_sync: false,
-        database_path: Some(PathBuf::from("sqlite::memory:")),
-        l1_provider_args: L1ProviderArgs::default(),
-        engine_api_url: None,
-        sequencer_args: SequencerArgs { sequencer_enabled: false, ..SequencerArgs::default() },
-        beacon_provider_args: BeaconProviderArgs::default(),
-        l2_provider_args: L2ProviderArgs::default(),
-    };
-    let node = ScrollRollupNode::new(node_args);
-    let NodeHandle { node, node_exit_future: _ } =
-        NodeBuilder::new(node_config.clone()).testing_node(exec.clone()).launch_node(node).await?;
-    let peer_id = *node.network.peer_id();
-    let node = NodeTestContext::new(node, scroll_payload_attributes).await?;
-
-    Ok((node, tasks, peer_id))
-}
-
-/// Helper function to create a new eth payload attributes
-fn scroll_payload_attributes(_timestamp: u64) -> ScrollPayloadBuilderAttributes {
-    let attributes = ScrollPayloadAttributes::default();
-    ScrollPayloadBuilderAttributes::try_new(B256::ZERO, attributes, 0).unwrap()
 }
